requirements.txt
appdirs==1.4.4
backcall==0.2.0
black==21.7b0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
entrypoints==0.3
et-xmlfile==1.1.0
ipykernel==6.3.0
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.2
jupyter-core==4.7.1
kiwisolver==1.3.2
matplotlib==3.4.3
matplotlib-inline==0.1.2
MedPy==0.4.0
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numpy==1.21.2
openpyxl==3.0.7
pandas==1.2.5
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
pydicom==2.2.0
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywin32==301
pyzmq==22.2.1
regex==2021.8.28
scipy==1.7.1
SimpleITK==2.1.0
six==1.16.0
tomli==1.2.1
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
wcwidth==0.2.5
XlsxWriter==3.0.1
